removing check on urn
